XLS support for SIMD registers.
Adds a simd field to Abi, plus code to have allow & prefer
sets for each interval.  We try to allocate from the prefer
set, but always at least take from the allow set.  Doubles
always prefer simd.  In addition, instructions that load and
store Cells, prefer simd.  This mimics the existing policy in
linear-scan.cpp.

Also changes linear-scan.cpp to not prefer GPRs for a tmp
that is an XMM candidate but which crosses a native call,
when HHIREnableCalleeSavedOpt==false (since prefering GPRs
in that case is part of the optimization).

Add support for XMM registers to PhysRegSaver.  If we had
a live XMM (double or TypedValue) that spanned a call, it
wasn't getting saved.

Reviewed By: @ottoni

Differential Revision: D1091845

+++ hphp/runtime/vm/jit/phys-reg.h
@@ -124,7 +124,7 @@ constexpr PhysReg InvalidReg;
  */
 struct RegSet {
   explicit RegSet() : m_bits(0) {}
-  explicit RegSet(PhysReg pr) : m_bits(1 << int(pr)) {}
+  explicit RegSet(PhysReg pr) : m_bits(uint64_t(1) << int(pr)) {}
 
   // Union
   RegSet operator|(const RegSet& rhs) const {
